Add support for custom event formatters
A custom event formatter is a function that takes a calendar event
object and returns a string from it. In order to be called, the function
must be stored at window.roamJsGcalCustomEventFormatter, which can be
done from any active Javascript block in Roam!

When the format string in the configuration contains the tag {custom},
then the custom formatter will be called if present, otherwise the tag
will not be replaced and will come out in the event text.